requirements.txt
python-telegram-bot
aiohttp
aiofiles
m3u8
ffmpeg-python
python-dotenv